move duplicate functions to one one
